🧪 Create helper function for validation tokens
